+++ src/org/jcvi/common/core/seq/read/trace/sanger/DefaultPositionSequence.java
@@ -22,8 +22,8 @@ public class DefaultPositionSequence implements PositionSequence{
 	}
 
 	@Override
-	public Position get(int index) {
-		return Position.valueOf(IOUtil.toUnsignedShort(data[index]));
+	public Position get(long index) {
+		return Position.valueOf(IOUtil.toUnsignedShort(data[(int)index]));
 	}
 
 	@Override
